[scripts] Updated pack_component to include pdbs as well

+++ scripts/pack_component.bat
@@ -19,10 +19,12 @@ if '%2'=='--debug' (
 set COMPONENT_DIR_NO_SLASH=%ROOT_DIR%component
 set RESULT_CONFIGURATION_DIR=%ROOT_DIR%_result\Win32_%CONFIGURATION%\
 set COMPONENT_DLL=%RESULT_CONFIGURATION_DIR%\bin\foo_spider_monkey_panel.dll
+set COMPONENT_PDB=%RESULT_CONFIGURATION_DIR%\dbginfo\foo_spider_monkey_panel.pdb
 set SAMPLES_COMPLETE_DIR_NO_SLASH=%ROOT_DIR%submodules\smp_2003
 set MOZ_JS_BIN_DIR=%ROOT_DIR%mozjs\%CONFIGURATION%\bin\
 set COMPONENT_OUT_DIR_NO_SLASH=%RESULT_CONFIGURATION_DIR%component
 set COMPONENT_OUT_DIR=%COMPONENT_OUT_DIR_NO_SLASH%\
+set COMPONENT_PDB_PACKAGE=%RESULT_CONFIGURATION_DIR%foo_spider_monkey_panel_pdb.zip
 set FB2K_ARCHIVE=%RESULT_CONFIGURATION_DIR%foo_spider_monkey_panel.fb2k-component
 
 echo Packing component to .fb2k-component
@@ -36,11 +38,32 @@ xcopy /r/y/s/q "%MOZ_JS_BIN_DIR%*.dll" "%COMPONENT_OUT_DIR%"
 if errorlevel 1 goto fail
 xcopy /r/y/q "%COMPONENT_DLL%" "%COMPONENT_OUT_DIR%"
 if errorlevel 1 goto fail
+if '%CONFIGURATION%'=='Debug' (
+    rem Only debug package should have pdbs inside
+    xcopy /r/y/s/q "%MOZ_JS_BIN_DIR%*.pdb" "%COMPONENT_OUT_DIR%"
+    if errorlevel 1 goto fail
+    xcopy /r/y/q "%COMPONENT_PDB%" "%COMPONENT_OUT_DIR%"
+    if errorlevel 1 goto fail
+) else (
+    rem Release pdbs are packed in a separate package
+    if exist "%COMPONENT_PDB_PACKAGE%" del /f/q "%COMPONENT_PDB_PACKAGE%"
+    7z a -tzip -mx=9 "%COMPONENT_PDB_PACKAGE%" "%MOZ_JS_BIN_DIR%*.pdb" > NUL
+    if errorlevel 1 goto fail
+    7z a -tzip -mx=9 "%COMPONENT_PDB_PACKAGE%" "%COMPONENT_PDB%" > NUL
+    if errorlevel 1 goto fail
+)
 
 if exist "%FB2K_ARCHIVE%" del /f/q "%FB2K_ARCHIVE%"
-7z a -tzip "%FB2K_ARCHIVE%" "%COMPONENT_OUT_DIR%*" > NUL
-echo Component was sucessfuly packed: %FB2K_ARCHIVE%
+7z a -tzip -mx=9 "%FB2K_ARCHIVE%" "%COMPONENT_OUT_DIR%*" > NUL
 if errorlevel 1 goto fail
+if '%CONFIGURATION%'=='Debug' (
+    echo Component was sucessfuly packed: %FB2K_ARCHIVE%
+) else (
+    echo Pdb's were sucessfuly packed: %COMPONENT_PDB_PACKAGE%
+    echo Component was sucessfuly packed: %FB2K_ARCHIVE%
+)
+
+
 exit /b 0
 
 :fail
